requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
databricks-sdk>=0.38.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
PyGithub>=2.5.0
pyyaml>=6.0.2
psycopg2-binary>=2.9.10
jinja2>=3.1.5
python-multipart>=0.0.18
httpx>=0.28.0
